pyproject.toml
[project]
name = "python-ai-playground"
version = "0.1.0"
description = "Learning playground for agentic AI: LangGraph, RAG, tool calling, vector DBs, code search & generation"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "anthropic>=0.39.0",
    "langchain-anthropic>=0.3.0",
    "langchain-core>=0.3.0",
    "langgraph>=0.2.0",
    "python-dotenv>=1.0.0",
    "PyGithub>=2.4.0",
]

[project.optional-dependencies]
# Weeks 1–2: RAG, vector DBs, “answer from docs”
rag = [
    "langchain>=0.3.0",
    "chromadb>=0.5.0",
    "langchain-chroma>=0.1.0",
]

# Weeks 3–4: Multi-agent, tool calling, “classify → search → respond”
agent = [
    "langgraph[anthropic]>=0.2.0",
]

# Weeks 5–6: Code embeddings, GitHub API, repo search
code-search = [
    "PyGithub>=2.4.0",
    "tiktoken>=0.8.0",
]

# Weeks 7–8: Code generation, validation, “detect bug → fix PR”
code-gen = [
    "ruff>=0.8.0",
    "httpx>=0.28.0",
]

# Install everything for the full learning path
all = [
    "langchain>=0.3.0",
    "chromadb>=0.5.0",
    "langchain-chroma>=0.1.0",
    "PyGithub>=2.4.0",
    "tiktoken>=0.8.0",
    "ruff>=0.8.0",
    "httpx>=0.28.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.8.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
